Added the SendNotifier interface to support CTP + LPL integration.
Create and drop a module into your system that listens for the SendNotifier events for the AM types related to CTP or Dissemination, then have it call LowPowerListening.setRxSleepInterval(msg, 512);.  Or capture the event for every AM type.

This can also be used to control other transmit parameters on a system-wide basis, like full duplex PA control on a per-destination basis, etc.  If this hook works out, we can expand this into other radio stacks.


git-svn-id: https://tinyos-main.googlecode.com/svn/trunk@3756 285f258d-8064-b571-fbc6-9656dc998e90