Sam3s DAC PDC transfer now uses PWM for sampling signal generation.
One caveat for now: the sampling signal has to be twice as one would expect. I don't know why. All the registers seem right.

For future improvement: the frequency calculations should be all done in Hz. Right now it is a mix of kHz and Hz. This looses accuracy in frequency generation.

git-svn-id: https://tinyos-main.googlecode.com/svn/trunk@5422 285f258d-8064-b571-fbc6-9656dc998e90